Adds dry run and confirmation for deletion

+++ lib/tasks/user.rake
@@ -5,7 +5,9 @@ namespace :user do
       raise 'Please specify the username of the user to be deleted' if args[:username].blank?
       
       user = User.find(username: args[:username])
-      raise "The username '#{args[:username]}' does not correspond to any user"
+      raise "The username '#{args[:username]}' does not correspond to any user" if user.nil?
+
+      raise 'Deletion aborted due to bad confirmation' if !deletion_confirmed?(user)
 
       user.destroy
     end
@@ -15,9 +17,28 @@ namespace :user do
       raise 'Please specify the email of the user to be deleted' if args[:email].blank?
       
       user = User.find(email: args[:email])
-      raise "The email '#{args[:email]}' does not correspond to any user"
+      raise "The email '#{args[:email]}' does not correspond to any user" if user.nil?
+
+      raise 'Deletion aborted due to bad confirmation' if !deletion_confirmed?(user)
 
       user.destroy
+    end
+
+    def deletion_confirmed?(user)
+      puts ""
+      puts "You are about to delete the following user:"
+      puts "\t> username: #{user.username}"
+      puts "\t> email: #{user.email}"
+      puts ""
+      puts "Alongside '#{user.username}', the following will be deleted:"
+      puts "\t> #{user.maps.length} maps"
+      puts "\t> #{user.tables.count} datasets"
+      puts ""
+      puts "Type in the user's username (#{user.username}) to confirm:"
+
+      confirm = STDIN.gets.strip
+      
+      confirm == user.username
     end
   end
 end
